requirements.txt
streamlit
google-generativeai
python-docx
streamlit
requests
python-docx
PyPDF2